feat: added category_detail, category_detail route
User can now select a category in the category list and be
redirected to the category detail page that will show all
the detail for the category. At this moment it's just a quick
description of the category. I expect it to show all items in the
category at some point.