feat: added item detail page + category redirect
When searching through the item list, the user can now click
on the item to be redirected to the item detail page. Here
they will see all the detail of the item. There is also a category
that the item belongs to that will also redirect the user to that specific
category detail page.