feat: added create new item form + POST to DB
User can now create a new item using the form provided. New item
will be posted to the DB and updated. User will then be redirected
to the item list to see their new item. We still need to do some
validation and sanitization but overall it's working as intended.